+++ test/legacy/ut/unified_platform/resource/ccu_transport/ut_ccu_connection.cc
@@ -60,7 +60,7 @@ protected:
 
 pair<unique_ptr<CcuConnection>, vector<unique_ptr<CcuJetty>>> MockMakeCcuConnection(TpProtocol tpProtocol)
 {
-    MOCKER(HrtGetDevicePhyIdByIndex).stubs().with(any()).will(returnValue(MAX_MODULE_DEVICE_NUM - 1));
+    MOCKER(HrtGetDevicePhyIdByIndex).stubs().with(any()).will(returnValue(static_cast<s32>(MAX_MODULE_DEVICE_NUM - 1)));
     
     HcclResult OkResult = HcclResult::HCCL_SUCCESS;
     HcclResult AgainResult = HcclResult::HCCL_E_AGAIN;
